Improve GPS update freq.
Standard we check the GPS every 60sec but when we get close to our
destination we should check the GPS pos every second to no miss the
final destination. By having 1800meters as the external zone we will
insure not to miss the final destination. So when we are 1800 form final
destination we set GPS freq to check every second until destination is
reached.